create lax_parallel and move parallel rules and primitives to enable a valid module import structure
