+++ jax/interpreters/parallel.py
@@ -24,6 +24,7 @@ from six.moves import reduce
 
 from .. import core
 from .. import linear_util as lu
+from .. import lax_parallel as lp
 from ..core import Trace, Tracer, Primitive, new_master
 from ..abstract_arrays import ShapedArray, ConcreteArray, make_shaped_array
 from ..util import safe_zip, unzip2, unzip3, partialmethod, prod
@@ -161,72 +162,8 @@ class PmapTrace(Trace):
     return PmapTracer(self, name, vals, axis)
 
 
-def unbound_name_error(primitive_name, *args, **kwargs):
-  axis_name = kwargs['axis_name']
-  msg = "axis name '{}' is unbound for primitive {}."
-  raise NameError(msg.format(axis_name, primitive_name))
-
-def PmapPrimitive(name):
-  prim = Primitive(name)
-  prim.def_impl(partial(unbound_name_error, name))
-  prim.def_abstract_eval(lambda x, *args, **kwargs: x)  # default
-  return prim
-
-
 pmap_primitive_rules = {}
 parallel_translation_rules = {}
-
-
-def psum(x, axis_name):
-  return psum_p.bind(x, axis_name=axis_name)
-
-def psum_pmap_rule(val, axis):
-  return val.sum(axis), None
-
-def psum_parallel_translation_rule(c, val, device_groups):
-  if len(device_groups) > 1:
-    return c.CrossReplicaSum(val, device_groups)
-  else:
-    return c.CrossReplicaSum(val)
-
-psum_p = PmapPrimitive('psum')
-pmap_primitive_rules[psum_p] = psum_pmap_rule
-parallel_translation_rules[psum_p] = psum_parallel_translation_rule
-
-
-def ptranspose(x, split_dim, concat_dim, **params):
-  return ptranspose_p.bind(x, split_dim=split_dim, concat_dim=concat_dim, **params)
-
-def ptranspose_shape_rule(x, split_dim, concat_dim, **params):
-  permutation = list(range(x.ndim))
-  permutation[concat_dim] = split_dim
-  permutation[split_dim] = concat_dim
-  return lax.transpose_shape_rule(x, permutation)
-
-def ptranspose_pmap_rule(x, axis, split_dim, concat_dim):
-  raise NotImplementedError
-
-def ptranspose_translation_rule(c, x, split_dim, concat_dim):
-  return c.AllToAll(x, split_dim, concat_dim)
-
-ptranspose_p = PmapPrimitive('ptranspose')
-ptranspose_p.def_abstract_eval(ptranspose_shape_rule)
-pmap_primitive_rules[ptranspose_p] = ptranspose_pmap_rule
-parallel_translation_rules[ptranspose_p] = ptranspose_translation_rule
-
-
-def all_gather(x, xdim, **params):
-  x = x.broadcast((xb.get_replica_count(),))
-  return ptranspose(x, 0, xdim, **params)
-
-def gather(x, axis_name):
-  return gather_p.bind(x, axis_name=axis_name)
-
-def gather_pmap_rule(val, axis):
-  return val, None
-
-gather_p = PmapPrimitive('gather')
-pmap_primitive_rules[gather_p] = gather_pmap_rule
 
 
 ### axis variable splitting and computation chunking
@@ -427,61 +364,21 @@ class PapplyTrace(Trace):
 papply_primitive_rules = {}
 
 
-def scatter_like(source, target):
-  return scatter_like_p.bind(source, target)
+pmap_primitive_rules[lp.psum_p] = lp.psum_pmap_rule
+parallel_translation_rules[lp.psum_p] = lp.psum_parallel_translation_rule
 
-def scatter_like_papply_rule(name, vals, axes):
-  source, target = vals
-  source_axis, target_axis = axes
-  assert source_axis is None
-  return _scatter(source, target, target_axis, name)
+pmap_primitive_rules[lp.gather_p] = lp.gather_pmap_rule
+papply_primitive_rules[lp.scatter_like_p] = lp.scatter_like_papply_rule
 
-scatter_like_p = Primitive('scatter_like')
-scatter_like_p.def_abstract_eval(lambda source, target: source)
-papply_primitive_rules[scatter_like_p] = scatter_like_papply_rule
-
+pmap_primitive_rules[lp.ptranspose_p] = lp.ptranspose_pmap_rule
+parallel_translation_rules[lp.ptranspose_p] = lp.ptranspose_translation_rule
 
 def defvectorized(prim):
-  papply_primitive_rules[prim] = partial(vectorized_papply, prim)
-
-def vectorized_papply(prim, name, vals, axes, **params):
-  assert all(axes[0] == a for a in axes[1:])
-  return prim.bind(*vals, **params), axes[0]
-
+  papply_primitive_rules[prim] = partial(lp.vectorized_papply, prim)
 
 def defreducer(prim, collective_prim):
-  papply_primitive_rules[prim] = partial(reducer_papply, prim, collective_prim)
-
-def reducer_papply(prim, cprim, name, vals, papply_axes, input_shape, axes):
-  operand, = vals
-  papply_axis, = papply_axes
-
-  other_axes = [i for i in axes if i != papply_axis]
-  if other_axes:
-    result = prim.bind(operand, axes=other_axes, input_shape=input_shape)
-  else:
-    result = operand
-
-  if not axes or papply_axis in axes:
-    return cprim.bind(result, axis_name=name), None
-  else:
-    new_papply_axis = papply_axis - onp.sum(onp.less(other_axes, papply_axis))
-    return result, new_papply_axis
-
+  papply_primitive_rules[prim] = partial(lp.reducer_papply, prim,
+                                         collective_prim)
 
 def defbroadcasting(prim):
-  papply_primitive_rules[prim] = partial(broadcasting_papply, prim)
-
-def broadcasting_papply(prim, name, vals, axes, **params):
-  x, y = vals
-  xdim, ydim = axes
-
-  if xdim is None:
-    return prim.bind(x, y, **params), ydim
-  elif ydim is None:
-    return prim.bind(x, y, **params), xdim
-  elif xdim == ydim:
-    return prim.bind(x, y, **params), xdim
-  else:
-    x = ptranspose(x, xdim, ydim, axis_name=xdim)
-    return prim.bind(x, y, **params), ydim
+  papply_primitive_rules[prim] = partial(lp.broadcasting_papply, prim)
